Refactor master code
Move implementation of master systems into dedicated classes for better
organization.
