+++ packages/master/package.json
@@ -20,9 +20,7 @@
   },
   "dependencies": {
     "@clusterio/lib": "^2.0.0-alpha.5",
-    "body-parser": "^1.18.2",
     "compression": "^1.7.2",
-    "cookie-parser": "^1.4.3",
     "express": "^4.16.2",
     "express-fileupload": "1.2.0",
     "fs-extra": "^8.1.0",
