+++ test/package.json
@@ -9,7 +9,6 @@
     "@clusterio/lib": "2.0.0-alpha.5",
     "@clusterio/master": "2.0.0-alpha.5",
     "@clusterio/slave": "2.0.0-alpha.5",
-    "body-parser": "^1.18.2",
     "express": "^4.16.2",
     "fs-extra": "^8.1.0",
     "jsonwebtoken": "^8.2.1",
